+++ tests/scriptlets/trusted-set-cookie.test.js
@@ -44,7 +44,7 @@ test('Set cookie string', (assert) => {
 
     assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
     assert.strictEqual(document.cookie.includes(cName), true, 'Cookie name has been set');
-    assert.strictEqual(document.cookie.includes(encodeURIComponent(cValue)), true, 'Cookie value has been set');
+    assert.strictEqual(document.cookie.includes(cValue), true, 'Cookie value has been set');
     clearCookie(cName);
 });